Julia plot formatting
In module make_plots, set ticks to the inside of the plot and show them
at every axis. Replace x/ylabel by ax[:x/ylabel].
New option '0' for ix in function rd_data in module pyp to assign all
columns in the input file as y data.
New options for title and fontsizes in function lineplot in module pyp.
New arg 'ti' for an AbstractString holding the title. New kwargs
ti_offset, ax_offset, and leg_offset to define differences in the
fontsizes of the title, axes labels, and legend compared to the tick
number fontsize.