Scaling in line plots, better error plotting, fix
In pyp.lineplot, introducing scaling factors 'SF' for y data. To
acknowledge scaling in the y label, it will have to be included manually
as given in the doc string.
Introducing err options '3' and '4' to add shading of ± a fraction of
the y data or times a factor n or 1/n, respectively. Previous option 3
is redeclared as option 5.
FIX: Inserting missing 'using' in the try using Dierckx statement in
     module VectorMaths.