Updated rd_data and lineplot routines in pyp, fixes
In module pyp in function rd_data introduce to read in files with
different column lengths. Number of columns can be specified with kwarg
ncol, otherwise it is taken from jlheaders or the first non-comment
line. If the number is columns of the current line is smaller the number
of total columns, the remaining "first" or "last" columns are filled
with NaN's according to the kwarg "fill". 
In function lineplot, updated derivation of ymin/ymax to include the the
error range. Updated treatment
FIX: In function lineplot, scaling the errors by SF, if needed.