New function multivector_mean and module numerics
In module VectorMaths, new function multivector_mean to obtain a common
xdata array from any number of dataframes holding the x and y data of
different vectors/datasets, as well as the mean and sum of all y data.
The function returns a DataFrame with the common x data, adjusted y
data, the mean and the some either only for the common x range of all
DataFrames or over the whole x data range, where the mean is scaled
before a delayed beginned or after a premature end of vectors to assure
a smooth mean in the transition region. Adding several new private
functions for multivector_mean.
New module numerics.jl in jl.mod with some useful functions currently
missing in julia:
- int():     to transform a float or string into an integer, if the hold
             a valid integer
- dig2int(): to reunite digits to an integer that have previously been
             assembled in an array from an integer using digits