Fix in multivector_mean
In SR multivector_mean in module VectorMaths, include outer boundaries
in the output matrix, when calculating the mean over the whole
datarange. Allow negative x values. This is achieved by using -Inf and
Inf instead of 0 and maximum(maxs) as lower and upper boundary, when
calling get_xdata for the lower_xdata and upper_xdata, respecitively.